backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
sqlalchemy>=1.4.23,<1.5.0
alembic>=1.7.1,<1.8.0
psycopg2-binary>=2.9.1,<3.0.0
python-dotenv>=0.19.0,<0.20.0
email-validator>=1.1.3,<1.2.0
fastapi-mail>=1.2.0,<1.3.0
Pillow==10.1.0